lexpilot_streamlit_app/requirements.txt
altair
pandas
streamlit
transformers
torch
PyPDF2
python-docx
psutil
